chore: remove detectron from setup.py (#6)
Having a non-pip dependency prevents unstructured-inference from being installed as a package from pypi, so this PR removes detectron2 from the setup.py dependencies. Also removed the code to fail the install if torch isn't installed - torch will be installed with layoutparser, and will be present when the user builds Detectron2 from source.