feat: ocr when too many chars unrecognized (#8)
Certain PDF documents have text blocks within their layout that contain text, but the text contains / consists of unrecognized characters, represented in the text as (cid:n) where n is an integer.

Current logic accepts whatever text is present, only applying OCR in the case the layout text is None. This PR changes the logic to also apply OCR in the case that > 50% of the characters are unrecognized.