chore: dl default model from huggingface (#30)
We've now seen cases where dropbox links are blacklisted, which prevent them from downloading the default detectron2 model. This PR changes the default model retrieval method to use huggingface (which already hosts the weights and config for the model). The code is also updated so getting the default model and a non-default model uses more of the same code.