refactor: generic model (#38)
Refactor of model code to make it less dependent on detectron2.

Creates an abstract model class UnstructuredModel, with inference triggered by calling the object. The class must be subclassed for each individual model type to at least guide the inference input to the right method (e.g. the detect method for detectron 2 models).

Detectron specific code has been moved to unstructured_inference.models.detectron2.

The code still uses layoutparser objects for convenience in the layout code preparing the input, and the output is assumed to be a layoutparser Layout object.