build(deps): update detectron2 (#154)
I noticed an issue that basic operations in this library were causing import errors. This turned out to be caused by the new version of PIL completing deprecation of certain references (specifically PIL.Image.LINEAR) used in the pinned version of detectron2 that we've been using. Updating to the latest commit of detectron2 fixes the problem.