fix: handle potential errors from tesseract call
We've seen a 500 error in `unstructured-api` due to an uncaught TesseractError in the `entire_page`
path. I can't reproduce it, but we can at least add a try catch. The last fix was too aggessive,
which we're tracking [here][https://github.com/Unstructured-IO/unstructured/issues/1086], so we may
need to adjust this fix as well.