Fix PDFMiner bug (#253)
Issue:

In some cases, PDFMiner identifies an image document as a full page and
in other installations not. It is difficult to find out when PDFMiner
behaves in one way or another. In either case tested, the version is
`pdfminer.six v20221105`. The solution is to ignore any annotation
coming from Chipper in case the full page clearing code is activated.
Not sure if this is relevant to other models.

---------

Co-authored-by: Antonio Jimeno Yepes <antonio@unstructured.io>