Feat/improve chipper bounding boxes (#292)
Chipper bounding boxes largely exceed the area where the annotated text
is. This PR intends solving this problem.

There are several fixes applied:

* Improved cross attention processing. Maps are filtered at the head
level, which is normalised [0-1], this improves the bounding box
definition.
* The correlation between the token index and cross attention map has
been resolved. Before, with beam search size = 1 or 3 there were cases
in which the cross attention map did not match the token being
processed.

In addition, the empty areas of the bounding boxes have been cleaned and
the overlaps between bounding boxes are resolved by identifying the
largest margin that separates both bounding boxes, identified as the
largest gap without text either in the horizontal or vertical
directions.

Note: overlapping bounding boxes for child elements are not resolved for
now. In the case of a list with list items, the list element will be
affected by the overlapping resolution code.

---------

Co-authored-by: Antonio Jimeno Yepes <antonio@unstructured.io>